% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_local_factors.R
\name{find_local_factors}
\alias{find_local_factors}
\title{Find the most sparse rotation of the leading Principal Components of a t by n matrix X.}
\usage{
find_local_factors(X, r, Lambda0)
}
\arguments{
\item{X}{A (usually standardized) t by n matrix of observations.}

\item{r}{An integer denoting the number of factors in X.}

\item{Lambda0}{A matrix that represents an orthonormal basis of the loading space. If not supplied, PCA is used.}
}
\value{
Returns a list with the following components:
\itemize{
\item \code{Lambda0} (if not supplied) the principal component estimate of the loading matrix (orthonormal)
\item \code{Lambda_rotated} a matrix that is the rotation of the loading matrix that produces the smallest l1-norm.
\item \code{rotation_diagnostics} a list containing 3 components"
\itemize{
\item \code{R} the rotation matrix that when used to rotate \code{Lambda0} produces the smallest l1-norm.
\item \code{l1_norm} a vector of length \code{r} containing the value of the l1 norm each solution generates
\item \code{sol_frequency} a vector of length \code{r} containing the frequency in the initial grid of each solution
}
}
}
\description{
Find the most sparse rotation of the leading Principal Components of a t by n matrix X.
}
