% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_factors.R
\name{local_factors}
\alias{local_factors}
\title{Test whether local factors are present in a given dataset \code{X} and return the rotation of the loading matrix with the smallest l1-norm.}
\usage{
local_factors(X, r)
}
\arguments{
\item{X}{A (usually standardized) t by n matrix of observations.}

\item{r}{An integer denoting the number of factors in X.}
}
\value{
Returns a list with the following components:
\itemize{
\item \code{has_local_factors} a logical equal to \code{TRUE} if local factors are present
\item \code{Lambda0} the principal component estimate of the loading matrix
\item \code{Lambda_rotated} a matrix that is the rotation of the loading matrix that produces the smallest l1-norm.
\item \code{rotation_diagnostics} a list containing 3 components"
\itemize{
\item \code{R} the rotation matrix that when used to rotate \code{Lambda0} produces the smallest l1-norm.
\item \code{l1_norm} a vector of length \code{r} containing the value of the l1 norm each solution generates
\item \code{sol_frequency} a vector of length \code{r} containing the frequency in the initial grid of each solution
}
}
}
\description{
Test whether local factors are present in a given dataset \code{X} and return the rotation of the loading matrix with the smallest l1-norm.
}
