% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example data with four factors from the replication files of \href{https://simonfreyaldenhoven.github.io/papers/factor_rotation.pdf}{Freyaldenhoven et. al (2025)}.}
\format{
\subsection{\code{example_data}}{

A matrix with 224 rows and 207 columns.
}
}
\source{
Dataset available as a .mat file can be generated from local_factors.zip \url{https://simonfreyaldenhoven.github.io/software/}
}
\usage{
example_data
}
\description{
Example data with four factors from the replication files of \href{https://simonfreyaldenhoven.github.io/papers/factor_rotation.pdf}{Freyaldenhoven et. al (2025)}.
}
\keyword{datasets}
