% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example data with four factors from the replication files of Freyaldenhoven et. al (2023).}
\format{
\subsection{\code{example_data}}{

A matrix with 500 rows and 300 columns.
}
}
\source{
Dataset available as a .mat file can be found as "example1.mat" under local_factors.zip \url{https://simonfreyaldenhoven.github.io/software/}
}
\usage{
example_data
}
\description{
Example data with four factors from the replication files of Freyaldenhoven et. al (2023).
}
\keyword{datasets}
