% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l1rotation-package.R
\docType{package}
\name{l1rotation-package}
\alias{l1rotation}
\alias{l1rotation-package}
\title{l1rotation: Identify Loading Vectors under Sparsity in Factor Models}
\description{
Estimate loading vectors using the l1 criterion as proposed in Freyaldenhoven (2025) \url{https://simonfreyaldenhoven.github.io/papers/factor_rotation.pdf}. Given a data matrix, find the rotation of the loading matrix with the smallest l1-norm and/or test for the presence of local factors.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SimonFreyaldenhoven/l1rotation}
  \item Report bugs at \url{https://github.com/SimonFreyaldenhoven/l1rotation/issues}
}

}
\author{
\strong{Maintainer}: Ryan Kobler \email{kobleary@gmail.com}

Authors:
\itemize{
  \item Simon Freyaldenhoven \email{simon.freyaldenhoven@phil.frb.org} [copyright holder]
}

}
\keyword{internal}
