% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_local_factors.R
\name{test_local_factors}
\alias{test_local_factors}
\title{Test for local factors, as in \code{local_factors} with additional flexibility.}
\usage{
test_local_factors(X, r, Lambda = NULL, eig_X = NULL, alpha_gamma = 0.05)
}
\arguments{
\item{X}{A (usually standardized) t by n matrix of observations.}

\item{r}{An integer denoting the number of factors in X.}

\item{Lambda}{(optional) a matrix that represents a sparse basis of the loading space}

\item{eig_X}{(optional) a vector of eigenvalues of X'X/t in decreasing order (only used if X is empty)}

\item{alpha_gamma}{(optional) a numeric value (default = 0.05)

#' @returns Returns a list with the following components:
\itemize{
\item \code{has_local_factors} a logical equal to \code{TRUE} if local factors are present
\item \code{n_small} an integer denoting the number of small loadings in sparse rotation
\item \code{gamma_n} an integer denoting the critical value to compare \code{n_small} to.
\item \code{h_n} a number denoting the cutoffused to determine which loadings are small
\item \code{Lambda} rotation of PCs with smallest l1-norm
\item \code{rotation_diagnostics} a list containing 3 components"
\itemize{
\item \code{R} the rotation matrix that when used to rotate \code{Lambda0} produces the smallest l1-norm.
\item \code{l1_norm} a vector of length \code{r} containing the value of the l1 norm each solution generates
\item \code{sol_frequency} a vector of length \code{r} containing the frequency in the initial grid of each solution
}
}}
}
\description{
Test for local factors, as in \code{local_factors} with additional flexibility.
}
