% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_local_factors.R
\name{test_local_factors}
\alias{test_local_factors}
\title{Test for the presence of local factors, as in \code{\link[=local_factors]{local_factors()}}, with additional flexibility.}
\usage{
test_local_factors(X, r, Lambda = NULL, alpha_gamma = 0.05)
}
\arguments{
\item{X}{A (usually standardized) t by n matrix of observations.}

\item{r}{An integer denoting the number of factors in X.}

\item{Lambda}{(optional) Matrix that represents a sparse basis of the loading space.}

\item{alpha_gamma}{(optional) A numeric value that represents the tuning parameter (default = 0.05).}
}
\value{
Returns a list with the following components:
\itemize{
\item \code{has_local_factors} Logical equal to \code{TRUE} if local factors are present.
\item \code{n_small} Integer denoting the number of small loadings in sparse rotation.
\item \code{gamma_n} Integer denoting the critical value to compare \code{n_small} to.
\item \code{h_n} Number denoting the cutoff used to determine which loadings are small.
\item \code{Lambda} Matrix that is the rotation of the loading matrix that produces the smallest l1-norm.
}
}
\description{
Test for the presence of local factors, as in \code{\link[=local_factors]{local_factors()}}, with additional flexibility.
}
\examples{
# Minimal example with 4 factors, where X is a 500 by 300 matrix
r <- 4
M <- nrow(example_data)
n <- ncol(example_data)

# Find minimum rotation
rotation_result <- find_local_factors(X = example_data, r)

# Test if sparse basis has local factors
test_result <- test_local_factors(
   X = example_data,
   r = r,
   Lambda = rotation_result$Lambda_rotated,
   alpha_gamma = 0.05
)

test_result$has_local_factors

}
